# server/swagger.yml
openapi: "3.0.0"
info:
  title: Keeper
  version: 1.0.0

servers:
  - url: /api

paths:
  /health:
    get:
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /token:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /note:
    get:
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: dir
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notes'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    post:
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /note/{key}:
    get:
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    put:
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

components:
  schemas:
    Message:
      type: object
      required:
        - ok
        - message
      properties:
        ok:
          type: boolean
        message:
          type: string
    TokenRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
    TokenResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    Notes:
      type: object
      required:
        - notes
      properties:
        notes:
          type: array
          items:
            type: string
        nextPage:
          type: string
    Note:
      type: object
      required:
        - key
        - body
      properties:
        key:
          type: string
        body:
          type: string
        modified:
          type: string
          format: date-time